Renamed src/vis/vis and src/vis/vis-model to src/vis/vis-view and src/vis/vis
