+++ src/geo/map.js
@@ -58,6 +58,17 @@ var Map = Model.extend({
         zoom: attrs.zoom || this.defaults.zoom
       });
     }
+
+    // This method is declared here so that we can spyOn _.debounce
+    // in the tests that depend on this to work
+    this.reload = _.debounce(function (options) {
+      var instanceOptions = {
+        sourceLayerId: options && options.sourceLayerId,
+        forceFetch: options && options.forceFetch
+      };
+
+      this._windshaftMap.createInstance(instanceOptions);
+    }, this.RELOAD_DEBOUNCE_TIME);
   },
 
   // PUBLIC API METHODS
@@ -171,15 +182,6 @@ var Map = Model.extend({
   _onDataviewAdded: function (layerModel) {
     this.reload();
   },
-
-  reload: _.debounce(function (options) {
-    var instanceOptions = {
-      sourceLayerId: options && options.sourceLayerId,
-      forceFetch: options && options.forceFetch
-    };
-
-    this._windshaftMap.createInstance(instanceOptions);
-  }, this.RELOAD_DEBOUNCE_TIME),
 
   _updateAttributions: function () {
     var defaultCartoDBAttribution = this.defaults.attribution[0];
